Add an initial CompositeBVH implementation
